Fixes issue 399.
When the user clicks submit now, the system will only ever submit once no matter how many times the user clicks the button. Once the page is reloaded they can submit in the same way again.

Also a further improvement to issue 403 means the validator is no longer firing as the user types in the taxon cell on the species grid. It fires only once the field loses focus.